# openapi.yaml
openapi: 3.0.3
info:
  title: PDF Extraction Service
  description: >
    FastAPI-based service for extracting text and metadata from PDF files.
    Supports full-text extraction and page-indexed extraction.
  version: "1.0.0"
servers:
  - url: http://localhost:8000
paths:
  /healthz:
    get:
      summary: Health check
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
  /extract/text:
    post:
      summary: Extract full text & metadata
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF file
      responses:
        "200":
          description: Extracted text and metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  filename:
                    type: string
                  pages:
                    type: integer
                  title:
                    type: string
                    nullable: true
                  author:
                    type: string
                    nullable: true
                  subject:
                    type: string
                    nullable: true
                  keywords:
                    type: string
                    nullable: true
                  text:
                    type: string
        "400":
          description: Bad request (non-PDF file)
        "413":
          description: File too large
        "422":
          description: PDF parse failed
  /extract/text/pages:
    post:
      summary: Extract text by page indices (0-based)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                payload:
                  type: string
                  description: JSON string with page indices, e.g. {"pages":[0,2,5]}
                file:
                  type: string
                  format: binary
                  description: PDF file
      responses:
        "200":
          description: Map of page index to text
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
        "413":
          description: File too large
        "422":
          description: PDF parse failed